+++ 5th Iteration/0. Results/Eoin/model/2017.07.26/10.24.07_RandomForestRegressor.txt
@@ -1,0 +1,104 @@
+RandomForestRegressor 20170726-102407
+
+RandomForestRegressor: Cross Validation (10 Folds)
+	Train Mean RMSE: 26202.00 (+/-467.94)
+	Test Mean RMSE: 44284.97 (+/-2569.11)
+	Train Mean R2: 0.99683 (+/-0.00012)
+	Test Mean R2: 0.99092 (+/-0.00088)
+
+
+	RandomForestRegressor number test predictions within 1 hour -> Mean: 656.5/2004 (+/- 16.0)
+	RandomForestRegressor % test predictions within 1 hour: -> Mean: 32.76% (+/- 0.80%)
+	RandomForestRegressor number test predictions within 24 hours -> Mean: 1909.5/2004 (+/- 10.6)
+	RandomForestRegressor % test predictions within 24 hours -> Mean: 95.28% (+/- 0.53%)
+
+
+Feature Importances:
+0. "Concurrent_open_cases" (0.634055)
+1. "Days_left_QTR" (0.211269)
+2. "AuditDuration" (0.097471)
+3. "Seconds_left_month" (0.048105)
+4. "HoldDuration" (0.003832)
+5. "AmountinUSD" (0.000751)
+6. "HoldTypeName_Customer" (0.000300)
+7. "HoldTypeName_3rd Party" (0.000282)
+8. "Source_eAgreement (Phy)" (0.000243)
+9. "SubReason_Basic Enterprise Commitment" (0.000237)
+10. "IsSignature" (0.000201)
+11. "StatusReason_Problem Solved" (0.000200)
+12. "HoldCount" (0.000173)
+13. "Complexity" (0.000157)
+14. "CountrySource_australia" (0.000147)
+15. "StageName" (0.000147)
+16. "AssignedToGroup_Internal" (0.000146)
+17. "AssignedToGroup_Submitter (Contact)" (0.000146)
+18. "SubReason_Meteaop" (0.000114)
+19. "CountryProcessed_northamerica" (0.000114)
+20. "Queue_NAOC" (0.000104)
+21. "SalesLocation_northamerica" (0.000099)
+22. "SubReason_Zero Usage Order" (0.000083)
+23. "OLSRevenue" (0.000079)
+24. "Source_Soft Copy" (0.000077)
+25. "Source_Web" (0.000074)
+26. "Revenutype_Current Revenue" (0.000073)
+27. "SubReason_Manual Order Entry" (0.000068)
+28. "SubReason_Future Pricing Only CPS" (0.000068)
+29. "SalesLocation_europe" (0.000062)
+30. "Revenutype_Future Billing" (0.000061)
+31. "Revenutype_Non-revenue" (0.000059)
+32. "IsGovernment" (0.000058)
+33. "Queue_EOC" (0.000054)
+34. "ROCName_EOC" (0.000053)
+35. "CountrySource_europe" (0.000049)
+36. "SalesLocation_australia" (0.000048)
+37. "ROCName_APOC" (0.000047)
+38. "Revenutype_Credit / Rebill" (0.000047)
+39. "CountrySource_asia" (0.000046)
+40. "CountrySource_northamerica" (0.000045)
+41. "HoldTypeName_Internal" (0.000038)
+42. "Source_Hard Copy" (0.000037)
+43. "SalesLocation_asia" (0.000036)
+44. "CountryProcessed_asia" (0.000034)
+45. "CountryProcessed_europe" (0.000032)
+46. "IsSOXCase" (0.000032)
+47. "Revenutype_Revenue Unknown" (0.000031)
+48. "Numberofreactivations" (0.000031)
+49. "SalesLocation_southamerica" (0.000030)
+50. "Queue_APOC" (0.000030)
+51. "Priority" (0.000030)
+52. "SubReason_True Up" (0.000028)
+53. "IsMagnumCase" (0.000021)
+54. "SubReason_Electronic Order Pend / Reject" (0.000020)
+55. "Source_Manual" (0.000019)
+56. "CountryProcessed_australia" (0.000016)
+57. "CountrySource_southamerica" (0.000015)
+58. "CountryProcessed_southamerica" (0.000015)
+59. "Queue_LOC" (0.000014)
+60. "Revenutype_Future OTRRR without OLS" (0.000009)
+61. "sourcesystem_CLT" (0.000008)
+62. "CountryProcessed_other" (0.000005)
+63. "SubReason_Tax Exemption Order" (0.000005)
+64. "AssignedToGroup_CRMT" (0.000004)
+65. "StatusReason_Information Provided" (0.000004)
+66. "AssignedToGroup_Ops. Program Manager" (0.000003)
+67. "Source_eAgreement (Ele)" (0.000002)
+68. "Revenutype_Future OTRRR with OLS" (0.000002)
+69. "Queue_E&E" (0.000001)
+70. "Revenutype_Revenue Impacting Case / Pending Revenue" (0.000001)
+71. "Queue_Broken" (0.000001)
+72. "AssignedToGroup_Microsoft IT" (0.000001)
+73. "CountrySource_other" (0.000001)
+74. "sourcesystem_NMEC" (0.000000)
+75. "sourcesystem_Aplquest" (0.000000)
+76. "SubReason_P&H Electronic Order" (0.000000)
+77. "Revenutype_New Work Sold" (0.000000)
+78. "Source_Fax" (0.000000)
+79. "Revenutype_Disputed Revenue" (0.000000)
+80. "sourcesystem_Moritz Jürgensen" (0.000000)
+81. "sourcesystem_clt" (0.000000)
+82. "AssignedToGroup_BPO" (0.000000)
+83. "sourcesystem_web" (0.000000)
+84. "Queue_Other" (0.000000)
+85. "sourcesystem_Current Revenue" (0.000000)
+86. "sourcesystem_NEMEC" (0.000000)
+87. "sourcesystem_Web" (0.000000)
