+++ 5th Iteration/0. Results/Eoin/prepare_dataset/2017.07.26/09.44.31_parameters.txt
@@ -1,0 +1,39 @@
+user: Eoin
+raw_data: UCD_Data_20170623_1.xlsx
+raw_data_location: ../../../Data/
+input_file: _24_7
+file_name: _24_7
+output_file: _26_7
+file_location: ../../../Data/
+iteration: 5
+notes: Testing_days_to_month_qtr_end
+clean_Incident: y
+clean_AuditHistory: n
+clean_HoldActivity: n
+clean_PackageTriageEntry: n
+LinearRegression: y
+ElasticNet: y
+KernelRidge: n
+RandomForestRegressor: y
+MLPRegressor: n
+GradientBoostingRegressor: n
+xgboost: n
+log_of_y: n
+seed: 12345
+histogram: n
+resample: n
+n_estimators: 10
+n_samples: 50
+crossvalidation: 10
+delete_ticketNumber: y
+rerun_with_top_importances: n
+top_k_features: 50
+export_importances_csv: n
+append_HoldDuration: y
+append_AuditDuration: y
+grid_search: n
+grid_search_RMSE: n
+last_4_BDays: n
+workload: y
+delete_created_resolved: y
+minimum_data: n
